requirements.txt
themoviedb
google-generativeai